backend/pyproject.toml
[project]
name = "indika-backend"
version = "0.1.0"
description = "Backend for AI Document and Multimedia Q&A"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.2",
  "python-multipart>=0.0.9",
  "pypdf>=5.0.0",
  "openai-whisper>=20240930",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "pgvector>=0.3.6",
  "PyJWT>=2.9.0",
  "redis>=5.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.7.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=95"
